+++ frontend/src/pages/Peserta/PresensiPage.jsx
@@ -203,7 +203,7 @@ const PresensiPage = () => {
           </div>
           <div className="inline-flex items-center gap-1.5 px-3 py-1.5 rounded-xl bg-amber-50 border border-amber-200 text-amber-900 text-xs font-semibold self-start sm:self-auto shrink-0">
             <Calendar size={14} className="text-amber-600" />
-            <span>{new Date().toLocaleDateString('id-ID', { weekday: 'short', day: '2-digit', month: 'short', year: 'numeric' })}</span>
+            <span>{new Date().toLocaleDateString('id-ID', { weekday: 'long', day: '2-digit', month: 'short', year: 'numeric' })}</span>
           </div>
         </div>
